Modifications so that there are less round-ups errors due to the use of floats, and when there are some, the values obtained are not taken into account for the study of the results.

+++ const.ml
@@ -1,21 +1,21 @@
 open Arg
 
 type parameter = {
-  rmax : float;
+  rmax : int;
   tree_type : int;
   size_of_tree : int;
   number_of_speeds : int;
-  max_speed : float;
+  max_speed : int;
   regularity_speed : int;
   number_of_tests : int
  }
 
 let default = {
-  rmax = 10.;
+  rmax = 10;
   tree_type = 0;
   size_of_tree = 15;
   number_of_speeds = 4;
-  max_speed = 15.;
+  max_speed = 15;
   regularity_speed = 1;
   number_of_tests = 50
 }
